+++ src/test/java/de/codescape/jira/plugins/scrumpoker/action/HealthCheckActionTest.java
@@ -8,8 +8,10 @@ import com.atlassian.jira.web.HttpServletVariables;
 import com.atlassian.upm.api.license.PluginLicenseManager;
 import com.atlassian.upm.api.license.entity.PluginLicense;
 import com.atlassian.upm.api.util.Option;
+import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerProject;
 import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
 import de.codescape.jira.plugins.scrumpoker.service.EstimationFieldService;
+import de.codescape.jira.plugins.scrumpoker.service.ProjectSettingService;
 import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingService;
 import org.junit.Rule;
 import org.junit.Test;
@@ -17,10 +19,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
+import java.util.List;
 
 import static de.codescape.jira.plugins.scrumpoker.action.HealthCheckAction.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class HealthCheckActionTest {
@@ -40,6 +45,9 @@ public class HealthCheckActionTest {
 
     @Mock
     private EstimationFieldService estimationFieldService;
+
+    @Mock
+    private ProjectSettingService projectSettingService;
 
     @InjectMocks
     private HealthCheckAction healthCheckAction;
@@ -142,6 +150,7 @@ public class HealthCheckActionTest {
     public void shouldSignalUnsetEstimationFieldIfEstimationFieldIsNotSet() {
         when(scrumPokerSettingService.load()).thenReturn(globalSettings);
         when(globalSettings.getStoryPointField()).thenReturn(null);
+        when(globalSettings.isDefaultProjectActivation()).thenReturn(true);
         assertThat(healthCheckAction.getConfigurationResults(), hasItem(Configuration.ESTIMATION_FIELD_NOT_SET));
     }
 
@@ -150,7 +159,26 @@ public class HealthCheckActionTest {
         when(scrumPokerSettingService.load()).thenReturn(globalSettings);
         when(globalSettings.getStoryPointField()).thenReturn("something");
         when(estimationFieldService.findStoryPointField()).thenReturn(null);
+        when(globalSettings.isDefaultProjectActivation()).thenReturn(true);
         assertThat(healthCheckAction.getConfigurationResults(), hasItem(Configuration.ESTIMATION_FIELD_NOT_FOUND));
+    }
+
+    @Test
+    public void shouldSignalNoProjectEnabledWhenNeitherGloballyEnabledNorHavingAProjectExplicitlyEnabled() {
+        when(scrumPokerSettingService.load()).thenReturn(globalSettings);
+        when(globalSettings.getStoryPointField()).thenReturn("something");
+        when(estimationFieldService.findStoryPointField()).thenReturn(null);
+        when(globalSettings.isDefaultProjectActivation()).thenReturn(false);
+        expectNoProjectExplicitlyEnabled();
+        assertThat(healthCheckAction.getConfigurationResults(), hasItem(Configuration.ESTIMATION_FIELD_NOT_FOUND));
+    }
+
+    private void expectNoProjectExplicitlyEnabled() {
+        List<ScrumPokerProject> scrumPokerProjects = new ArrayList<>();
+        ScrumPokerProject scrumPokerProject = mock(ScrumPokerProject.class);
+        when(scrumPokerProject.isScrumPokerEnabled()).thenReturn(false);
+        scrumPokerProjects.add(scrumPokerProject);
+        when(projectSettingService.loadAll()).thenReturn(scrumPokerProjects);
     }
 
     @Test
@@ -158,6 +186,7 @@ public class HealthCheckActionTest {
         when(scrumPokerSettingService.load()).thenReturn(globalSettings);
         when(globalSettings.getStoryPointField()).thenReturn("something");
         when(estimationFieldService.findStoryPointField()).thenReturn(estimationField);
+        when(globalSettings.isDefaultProjectActivation()).thenReturn(true);
         assertThat(healthCheckAction.getConfigurationResults(), is(empty()));
     }